refactor: move transformer utils to another file
